delete image implemented

+++ app/src/main/java/com/chris/geminibasedapp/ui/screen/SavedTextGenerationChatScreen.kt
@@ -7,6 +7,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.BasicAlertDialog
 import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -17,6 +19,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
@@ -36,6 +41,14 @@ fun SavedTextGenerationChatScreen(
 
     val uiState by aiViewModel.uiState.collectAsState()
     val savedTextGenChat by aiViewModel.savedTextGenerationChatList.collectAsState()
+
+    var confirmDelete by remember {
+        mutableStateOf(false)
+    }
+
+    var selectedChat by remember {
+        mutableStateOf("")
+    }
 
     fun readList() {
         aiViewModel.readUserSavedChatList(
@@ -57,7 +70,6 @@ fun SavedTextGenerationChatScreen(
                     .fillMaxSize()
                     .padding(vertical = 32.dp, horizontal = 16.dp)
             ) {
-
                 items(savedTextGenChat) {chatTitle ->
                     ListCard(
 
@@ -66,15 +78,25 @@ fun SavedTextGenerationChatScreen(
                             onItemClick(chatTitle.id,chatTitle.title)
                         },
                         itemDelete = {
-                            aiViewModel.deleteSavedTextGenChat(
-                                user = aiViewModel.currentUser!!,
-                                chatTitle.id)
-                            readList()
+                            selectedChat = chatTitle.id
+                            confirmDelete = true
                         }
                     )
                 }
-
             }
+        }
+
+        if (confirmDelete) {
+            DeleteSavedChatDialog(
+                onDismiss = { confirmDelete = false},
+                onConfirm = {
+                    aiViewModel.deleteSavedTextGenChat(
+                        user = aiViewModel.currentUser!!,
+                        documentId = selectedChat
+                    )
+                    readList()
+                }
+                )
         }
 
 
@@ -113,4 +135,27 @@ fun ListCard(
 
         }
     }
+}
+
+@Composable
+fun DeleteSavedChatDialog(
+    onDismiss: () -> Unit,
+    onConfirm: () -> Unit
+) {
+    AlertDialog(
+        title = {
+            Text(text = "Delete Selected Chat")
+        },
+        text = {
+            Text(text = "Do you want to delete selected chat ?")
+        },
+        onDismissRequest = { onDismiss() },
+        confirmButton = {
+            IconButton(onClick = {
+                onConfirm()
+                onDismiss()
+            }) {
+                Icon(painter = painterResource(id = R.drawable.baseline_delete_24), contentDescription = "Delete")
+            }
+        })
 }
